Added table_fragments_util module.
