+++ src/common/src/util/mod.rs
@@ -45,3 +45,4 @@ pub use future_utils::{
 };
 #[macro_use]
 pub mod match_util;
+pub mod table_fragments_util;
