+++ src/rpc_client/src/meta_client.rs
@@ -34,6 +34,7 @@ use risingwave_common::util::addr::HostAddr;
 use risingwave_common::util::column_index_mapping::ColIndexMapping;
 use risingwave_common::util::resource_util::cpu::total_cpu_available;
 use risingwave_common::util::resource_util::memory::system_memory_available_bytes;
+use risingwave_common::util::table_fragments_util::downgrade_table_fragments;
 use risingwave_common::RW_VERSION;
 use risingwave_hummock_sdk::compaction_group::StateTableId;
 use risingwave_hummock_sdk::{
@@ -77,6 +78,7 @@ use risingwave_pb::meta::scale_service_client::ScaleServiceClient;
 use risingwave_pb::meta::serving_service_client::ServingServiceClient;
 use risingwave_pb::meta::stream_manager_service_client::StreamManagerServiceClient;
 use risingwave_pb::meta::system_params_service_client::SystemParamsServiceClient;
+use risingwave_pb::meta::table_fragments::GraphRenderType;
 use risingwave_pb::meta::telemetry_info_service_client::TelemetryInfoServiceClient;
 use risingwave_pb::meta::update_worker_node_schedulability_request::Schedulability;
 use risingwave_pb::meta::*;
@@ -844,7 +846,14 @@ impl MetaClient {
 
     pub async fn get_cluster_info(&self) -> Result<GetClusterInfoResponse> {
         let request = GetClusterInfoRequest {};
-        let resp = self.inner.get_cluster_info(request).await?;
+        let mut resp = self.inner.get_cluster_info(request).await?;
+
+        for table_fragments in &mut resp.table_fragments {
+            if table_fragments.graph_render_type == GraphRenderType::RenderTemplate as i32 {
+                downgrade_table_fragments(table_fragments);
+            }
+        }
+
         Ok(resp)
     }
